backend/requirements.txt
Flask
openai
supabase
feedparser
psycopg2-binary
python-dotenv
flask_cors